renamed entity Preferences
