+++ src/app/providers/TinyBaseProvider.tsx
@@ -4,7 +4,7 @@ import { Inspector } from "tinybase/ui-react-inspector";
 
 /* ---------- CODE BLOCK: tinyBase stores required to add them to <Provider> ---------- */
 import { JournalStore } from "@/entities/journal";
-import { PreferencesStore } from "@/entities/preferences";
+import { PreferencesStore } from "@/entities/user-config";
 import { RecordDraftStore } from "@/features/create-record/model/RecordDraftStore";
 
 type AppProps = {
