+++ src/entities/user-config/model/Preferences.ts
