+++ src/entities/user-config/model/preferencesSchema1.ts
